pip: split version logic into separe lib file; stashutils.wheels: handle extra and version dependend dependencies
